Add a working mobile menu for smaller screens (#27)
This PR adds the mobile first menu that works on extra small screens.
This is anything smaller than 640px wide. Once we hit the `sm` screen
size breakpoint according to the responsive design set by
[TailwindCSS](https://tailwindcss.com/docs/responsive-design) we'll
switch to using the NavigationMenu.

Once the menu gets a bit larger we'll need to add in scrolling
capability with the mobile menu content area.